backend/requirements.txt
fastapi
uvicorn
firebase-admin
pandas
numpy
openpyxl
python-multipart
python-dotenv
